manulife-backend/requirements.txt
fastapi
uvicorn
requests
azure-identity
urllib3
pydantic